requirements.txt
flask
flask-sqlalchemy
pytest
requests
allure-pytest